2/15 퇴근 시간 업데이트 수정 완료

+++ src/main/java/com/wittypuppy/backend/attendance/dto/EmployeeDTO.java
@@ -1,6 +1,5 @@
 package com.wittypuppy.backend.attendance.dto;
 
-import com.wittypuppy.backend.Employee.dto.User;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -10,9 +9,9 @@ import java.time.LocalDateTime;
 @Setter
 @ToString
 
-public class EmployeeDTO {
+public class EmployeeDTO  {
 
-    private Long employeeCode;
+    private int employeeCode;
 
     private DepartmentDTO departmentCode;
